Fix the armored jetpack not properly updating values based on config changes, and move the attribute caching helper stuff from MekanismTools to the main module so that we can make use of it in the armored jetpack
