+++ src/main/java/mekanism/common/item/gear/ItemFreeRunners.java
@@ -171,18 +171,8 @@ public class ItemFreeRunners extends ArmorItem implements ISpecialGear, IItemHUD
     protected static class FreeRunnerMaterial extends BaseSpecialArmorMaterial {
 
         @Override
-        public int getDamageReductionAmount(EquipmentSlotType slotType) {
-            return 0;
-        }
-
-        @Override
         public String getName() {
             return Mekanism.MODID + ":free_runners";
-        }
-
-        @Override
-        public float getToughness() {
-            return 0;
         }
     }
 }
